Moved some classes to files.
